Style login page
This adds more basic styling to the login page to make it look pretty.
In doing so, some basic SCSS infrastructure is added (mixins, tweaks to
core theme components, etc).

We'll want to find a place for the all the
styles, for now everything is just lumped into application.scss.

Finally, the option for adding content outside of the main page
container is added (this enables things like the login "banner")

Usage in a view:

  <% provide :before_container do %>
    <!-- this content will come before the main page container -->
  <% end %>

  <!-- this content will be in the main page container -->