Add Campaign model
This adds a basic Campaign model as outlined in the SRS. Currently, a
field is added to the user signup form for a pilot number that is used
to automatically create a Campaign. It's possible to have a separate
view, perhaps as part of some flow, but it would be nice if we could
avoid it.