Add sample notification bar
This adds a template for a notification bar that will only show up on
mobile, with an avatar / profile link, as well as a status switching
component and an indicator when a call comes in.

Components are non-functional - we'll probably want the status toggle to
fire an AJAX request to a controller to update the agent's status, and
slide down the menu and flash the indicator on the `call_start` event.

This commit also adds the Bootstrap glyphicons, which will be useful
elsewhere.